requirements.txt
Pillow>=8.1.1
PyYAML
easydict
numpy
tensorboardX>=2.0
